apps and models separated
