+++ back/identify/Github/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from . import models
+
+# Register your models here.
+
+admin.site.register(models.Github)
