+++ back/identify/Google/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from . import models
+
+
+# Register your models here.
+
+admin.site.register(models.Google)
