+++ back/identify/Google/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class GoogleConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'Google'
