+++ back/identify/LinkedIn/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+from . import models
+# Register your models here.
+
+admin.site.register(models.LinkedIn)
+
+admin.site.register(models.Company)
+
+admin.site.register(models.Skill)
